client: rename `client(config, servers)` of mobile
To have a unify flow to initialize VPN network, the funcion renamed as
the same cli using.

`mobile::lightway::async_lightway_start(..)` is moved to
`client(ClientConfig, Vec<ClientConnectionConfig>)` with the same
function name and function inputs.
