+++ lightway-client/src/lib.rs
@@ -16,21 +16,24 @@ pub mod state;
 #[cfg(feature = "mobile")]
 uniffi::setup_scaffolding!();
 
-use anyhow::Result;
 #[cfg(desktop)]
-use anyhow::{Context, anyhow};
+use anyhow::Context;
+use anyhow::{Result, anyhow};
 use bytes::BytesMut;
+#[cfg(desktop)]
 use bytesize::ByteSize;
+#[cfg(feature = "mobile")]
+use futures::stream::FusedStream;
 #[cfg(desktop)]
 use futures::{FutureExt, stream::FuturesUnordered};
 pub use io::inside::{InsideIO, InsideIORecv};
 use keepalive::Keepalive;
 use lightway_app_utils::EventStreamCallback;
-#[cfg(desktop)]
-use lightway_app_utils::PacketCodecFactoryType;
-use lightway_app_utils::{ConnectionTicker, ConnectionTickerState, TunConfig, args::Cipher};
+use lightway_app_utils::{ConnectionTicker, ConnectionTickerState, args::Cipher};
 #[cfg(desktop)]
 use lightway_app_utils::{DplpmtudTimer, connection_ticker_cb};
+#[cfg(desktop)]
+use lightway_app_utils::{PacketCodecFactoryType, TunConfig};
 #[cfg(desktop)]
 use lightway_core::{BuilderPredicates, ClientContextBuilder, EventCallback};
 use lightway_core::{
@@ -49,16 +52,21 @@ use crate::io::outside::BATCH_RECV_SIZE;
 use crate::keepalive::Config as KeepaliveConfig;
 #[cfg(desktop)]
 use crate::route_manager::{RouteManager, RouteMode};
+#[cfg(feature = "mobile")]
+use crate::state::DeviceNetworkState;
 #[cfg(feature = "debug")]
 use lightway_app_utils::wolfssl_tracing_callback;
 pub use lightway_core::{
     AuthMethod, MAX_INSIDE_MTU, MAX_OUTSIDE_MTU, PluginFactoryError, PluginFactoryList,
     RootCertificate, Version,
 };
+
 #[cfg(feature = "debug")]
 // re-export so client app does not need to depend on lightway-core
 pub use lightway_core::{enable_tls_debug, set_logging_callback};
 use pnet_packet::ipv4::Ipv4Packet;
+#[cfg(feature = "mobile")]
+use std::collections::HashMap;
 #[cfg(desktop)]
 use std::future::Future;
 #[cfg(desktop)]
@@ -71,16 +79,19 @@ use std::{
     sync::{Arc, Mutex, Weak},
     time::Duration,
 };
-#[cfg(desktop)]
-use tokio::task::{JoinHandle, JoinSet};
+use tokio::task::JoinHandle;
 use tokio::{
     net::{TcpStream, UdpSocket},
-    sync::{mpsc, oneshot},
+    sync::oneshot,
 };
 #[cfg(desktop)]
-use tokio_stream::{StreamExt, StreamMap};
+use tokio::{sync::mpsc, task::JoinSet};
+use tokio_stream::StreamExt;
 #[cfg(desktop)]
-use tracing::info;
+use tokio_stream::StreamMap;
+#[cfg(feature = "mobile")]
+use tracing::{Instrument, debug, info_span, warn};
+use tracing::{error, info};
 
 /// Connection type
 /// Applications can also attach socket for library to use directly,
@@ -139,6 +150,7 @@ pub enum LightwayError {
     StateError(#[from] crate::state::Error),
 }
 
+#[cfg(not(feature = "mobile"))]
 #[derive(educe::Educe)]
 #[educe(Debug)]
 pub struct ClientConfig<ExtAppState: Send + Sync> {
@@ -245,6 +257,17 @@ pub struct ClientConfig<ExtAppState: Send + Sync> {
     /// File path to save wireshark keylog
     #[cfg(feature = "debug")]
     pub keylog: Option<PathBuf>,
+}
+
+#[cfg(feature = "debug")]
+pub struct ClientConfig {
+    tun_local_ip: Ipv4Addr,
+    tun_dns_ip: Ipv4Addr,
+    tun_fd: std::os::fd::RawFd,
+    preferred_connection_wait_interval: std::time::Duration,
+
+    external_event_handler: Arc<dyn crate::event_handlers::EventHandlers>,
+    connected_index: Arc<std::sync::OnceLock<usize>>,
 }
 
 /// The Conection configures for desktop clients
@@ -543,7 +566,7 @@ pub async fn inside_io_task<ExtAppState: Send + Sync>(
     }
 }
 
-#[cfg(desktop)]
+#[cfg(not(feature = "mobile"))]
 async fn handle_network_change<ExtAppState: Send + Sync>(
     keepalive: Keepalive,
     mut network_change_signal: mpsc::Receiver<()>,
@@ -567,6 +590,73 @@ async fn handle_network_change<ExtAppState: Send + Sync>(
         }
     }
     ClientResult::UserDisconnect
+}
+
+#[cfg(feature = "mobile")]
+/// Handle all the network changes on the device.
+async fn handle_network_change(
+    keepalive: Keepalive,
+    mut network_change_receiver: tokio::sync::mpsc::Receiver<DeviceNetworkState>,
+    weak: Weak<Mutex<Connection<ConnectionState<mobile::lightway::TunnelState>>>>,
+    #[cfg_attr(not(apple), allow(unused_variables))] reset_outside_io_tx: tokio::sync::mpsc::Sender<
+        (),
+    >,
+) -> uniffi::Result<ClientResult> {
+    while let Some(network_state) = network_change_receiver.recv().await {
+        let Some(conn) = weak.upgrade() else {
+            return Ok(ClientResult::UserDisconnect);
+        };
+        let conn_type = conn.lock().unwrap().connection_type();
+        info!("Device network state change to {network_state:?}");
+        match network_state {
+            DeviceNetworkState::Online | DeviceNetworkState::InterfaceChanged => {
+                match conn_type {
+                    ConnectionType::Datagram => {
+                        // We only need to use reset new socket on iOS but not on Android
+                        #[cfg(apple)]
+                        {
+                            // Reset UDP transport when the network changes. This will ensure the udp traffic is routed
+                            // via the new network path and avoid using mobile data unnecessarily.
+                            info!("resetting udp transport due to network change ..");
+                            reset_outside_io_tx.send(()).await?;
+                        }
+                        #[cfg(not(apple))]
+                        // Trigger a new keepalive on the new socket (keepalive will
+                        // be triggered after the socket has been updated on iOS)
+                        keepalive.network_changed().await;
+                    }
+                    ConnectionType::Stream => {
+                        info!("client shutting down due to network change ..");
+                        let _ = conn.lock().unwrap().disconnect();
+                        return Ok(ClientResult::NetworkChange);
+                    }
+                }
+            }
+            DeviceNetworkState::RouteUpdated => {
+                match conn_type {
+                    ConnectionType::Datagram => {
+                        // UDP survives updating the network, but to ensure that the new network
+                        // with the current endpoint works reliably, we trigger keep-alives
+                        // e.g swapping from an unrestricted network to one that blocks UDP
+                        info!("sending keepalives due to network change ..");
+                        keepalive.network_changed().await;
+                    }
+                    ConnectionType::Stream => {
+                        info!("client shutting down due to network change ..");
+                        let _ = conn.lock().unwrap().disconnect();
+                        return Ok(ClientResult::NetworkChange);
+                    }
+                }
+            }
+            DeviceNetworkState::Offline => {
+                // Suspend keepalive timers as we're guaranteed
+                // to fail and disconnect after all failed attempts
+                info!("suspending keepalive...");
+                keepalive.suspend().await;
+            }
+        }
+    }
+    Ok(ClientResult::UserDisconnect)
 }
 
 pub async fn handle_encoded_pkt_send<ExtAppState: Send + Sync>(
@@ -630,7 +720,7 @@ pub async fn handle_decoded_pkt_send<ExtAppState: Send + Sync>(
             if err.is_fatal(conn.connection_type()) {
                 return Err(err.into());
             }
-            tracing::error!("Failed to process outside data: {err}");
+            error!("Failed to process outside data: {err}");
         }
     }
 
@@ -648,7 +738,7 @@ pub async fn encoding_request_task<ExtAppState: Send + Sync>(
         };
 
         if let Err(e) = conn.lock().unwrap().set_encoding(enable) {
-            tracing::error!(
+            error!(
                 "Error encoutered when trying to send encoding request. {}",
                 e
             );
@@ -760,7 +850,7 @@ pub async fn connect<
                 #[cfg_attr(not(batch_receive), allow(unused_mut))]
                 let mut sock = io::outside::Udp::new(server_config.server, maybe_sock)
                     .await
-                    .inspect_err(|e| tracing::error!("Failed to create outside IO UDP socket: {e}"))
+                    .inspect_err(|e| error!("Failed to create outside IO UDP socket: {e}"))
                     .context("Outside IO UDP")?;
 
                 #[cfg(batch_receive)]
@@ -773,7 +863,7 @@ pub async fn connect<
             ClientConnectionMode::Stream(maybe_sock) => {
                 let sock = io::outside::Tcp::new(server_config.server, maybe_sock)
                     .await
-                    .inspect_err(|e| tracing::error!("Failed to create outside IO TCP socket: {e}"))
+                    .inspect_err(|e| error!("Failed to create outside IO TCP socket: {e}"))
                     .context("Outside IO TCP")?;
                 (ConnectionType::Stream, Arc::new(sock))
             }
@@ -1147,7 +1237,7 @@ fn validate_client_config<
 /// priority connection (in the specified array order).
 ///
 /// stop_signal sends a signal if the program received INT/TERM signals
-#[cfg(desktop)]
+#[cfg(not(feature = "mobile"))]
 pub async fn client<
     EventHandler: 'static + Send + EventCallback,
     ExtAppState: 'static + Default + Send + Sync,
@@ -1199,11 +1289,11 @@ pub async fn client<
         let connect_futs: FuturesUnordered<_> = conn_confs
             .into_iter()
             .enumerate()
-            .map(|(index, server_config)| {
+            .map(|(index, conn_config)| {
                 let config = &config;
                 let inside_io = inside_io.clone();
                 async move {
-                    let result = connect(config, server_config, inside_io)
+                    let result = connect(config, conn_config, inside_io)
                         .await
                         .map(|mut conn| (conn.connected_signal.take().unwrap(), conn));
                     (index, result)
@@ -1231,7 +1321,7 @@ pub async fn client<
     if let Some(signal) = config.best_connection_selected_signal.take()
         && signal.send(best_connection_index).is_err()
     {
-        tracing::error!("Failed to send best_connection_selected_signal");
+        error!("Failed to send best_connection_selected_signal");
     }
 
     let pos = connections
@@ -1249,7 +1339,7 @@ pub async fn client<
         tokio::spawn(async move {
             while network_change_signal.recv().await.is_some() {
                 if let Err(e) = connection_network_change_signal.send(()).await {
-                    tracing::error!("Failed to send network_change_signal: {e}");
+                    error!("Failed to send network_change_signal: {e}");
                 }
             }
         });
@@ -1260,7 +1350,7 @@ pub async fn client<
         tokio::spawn(async move {
             while let Some(enabled) = inside_pkt_codec_config.encoding_request_signal.recv().await {
                 if let Err(e) = connection_encoding_request_signal.send(enabled).await {
-                    tracing::error!("Failed to send encoding_request_signal: {e}");
+                    error!("Failed to send encoding_request_signal: {e}");
                 }
             }
         });
@@ -1270,7 +1360,7 @@ pub async fn client<
     tokio::spawn(async move {
         let _ = stop_signal.await;
         if let Err(()) = connection_stop_signal.send(()) {
-            tracing::error!("Failed to send stop signal");
+            error!("Failed to send stop signal");
         }
     });
 
@@ -1296,6 +1386,266 @@ pub async fn client<
     }
 
     result
+}
+
+#[cfg(feature = "mobile")]
+pub(crate) async fn client(
+    crate::ClientConfig {
+        tun_local_ip,
+        tun_dns_ip,
+        tun_fd,
+        preferred_connection_wait_interval,
+
+        external_event_handler,
+        connected_index,
+    }: crate::ClientConfig,
+    conn_confs: Vec<crate::config::ConnectionConfig>,
+) -> uniffi::Result<ClientResult> {
+    let mut outside_sockets = conn_confs
+        .iter()
+        .map(|s| {
+            crate::OutsideSocket::new(s.mode.is_tcp(), Some(external_event_handler.clone())).ok()
+        })
+        .collect::<Vec<Option<crate::OutsideSocket>>>();
+
+    // Strore meta data before the server consumed
+    let server_len = conn_confs.len();
+    let tcp_connections_only = conn_confs.iter().all(|s| s.mode.is_tcp());
+
+    let inside_io =
+        mobile::lightway::setup_tunnel_interface(tun_fd, tun_local_ip, tun_dns_ip).await?;
+
+    let (_network_change_sender, mut network_change_receiver) = tokio::sync::mpsc::channel(1);
+
+    let (online_signal_sender, mut online_signal) = tokio::sync::mpsc::channel(server_len);
+    let (event_handler, stream) = EventStreamCallback::new();
+    tokio::spawn(event_handlers::handle_global_events(
+        stream,
+        tokio::time::Instant::now(),
+        external_event_handler.clone(),
+    ));
+    let mut client_connect_configs = Vec::new();
+    for (instance_id, conn_conf) in conn_confs.into_iter().enumerate() {
+        let client_connect_config = conn_conf.into_client_connection_config(
+            instance_id,
+            &mut outside_sockets,
+            online_signal_sender.clone(),
+            event_handler.clone(),
+            external_event_handler.clone(),
+        )?;
+        client_connect_configs.push(client_connect_config)
+    }
+
+    let (mut in_progress_connection_abort_handles, mut in_progress_connections): (
+        Vec<_>,
+        futures::stream::FuturesUnordered<_>,
+    ) = client_connect_configs
+        .into_iter()
+        .map(|c| {
+            let instance_id = c.instance_id;
+            let task = tokio::spawn(
+                mobile::lightway::lightway_client_connect(c)
+                    .instrument(info_span!("LightwayConnection", instance_id)),
+            );
+            (task.abort_handle(), task)
+        })
+        .unzip();
+
+    let mut wait_timer_task = tokio::spawn(tokio::time::sleep(preferred_connection_wait_interval));
+
+    debug!(
+        "Creating {} parallel connections",
+        in_progress_connections.len()
+    );
+
+    drop(outside_sockets);
+    drop(event_handler);
+
+    // Drop the last sender
+    drop(online_signal_sender);
+
+    let mut non_preferred_connections: Vec<(usize, mobile::lightway::LightwayConnection)> =
+        Vec::new();
+    let mut pending_online_connections: HashMap<usize, mobile::lightway::LightwayConnection> =
+        HashMap::with_capacity(server_len);
+    let mut failed_connections = 0usize;
+    let mut connection_error_to_return = None;
+
+    debug!("Waiting for online signal");
+    let active_connection = loop {
+        tokio::select! {
+            // Prioritise management commands over other branches, also make sure we add
+            // LightwayConnection to HashMap first to make sure when it goes online,
+            // we can remove it from the HashMap and break the loop.
+            biased;
+            _ = futures::future::ready(()), if failed_connections == server_len => {
+                error!("All connections failed, exiting...");
+                return Err(connection_error_to_return.unwrap_or(anyhow!("All connections failed")));
+            }
+
+            // On iOS specifically:
+            // If the library is called during a network change, it is possible all the TCP sockets
+            // created during this moment are "offline", and they cannot reach the servers.
+
+            // We early return and end connections attempt earlier for both Android and iOS as
+            // the connected server are going to get reset by the network change later anyway.
+            Some(DeviceNetworkState::Online | DeviceNetworkState::RouteUpdated | DeviceNetworkState::InterfaceChanged) = network_change_receiver.recv(), if tcp_connections_only => {
+                info!("client shutting down due to network change while connecting for Lightway - TCP");
+                return Ok(ClientResult::NetworkChange)
+            },
+
+            Some(connection_result) = in_progress_connections.next(), if !in_progress_connections.is_terminated() => {
+                match connection_result {
+                    Ok(Ok(connection)) => {
+                        debug!("Adding connections to hash map");
+                        let _ = pending_online_connections.insert(connection.instance_id, connection);
+                        continue;
+                    },
+                    Ok(Err(e)) => error!("Error while waiting for connection to set up: {:?}", e),
+                    Err(e) => error!("Join Error while waiting for connection to set up: {:?}", e)
+                };
+                failed_connections += 1;
+            },
+
+            _ = &mut wait_timer_task, if !wait_timer_task.is_finished() => {
+                if !non_preferred_connections.is_empty() {
+                    non_preferred_connections.sort_by_key(|(instance_id, _)| *instance_id);
+                    let (instance_id, connection) = non_preferred_connections.swap_remove(0);
+                    info!(?instance_id, "Defer timeout, choosing best connection");
+                    break connection;
+                }
+            },
+
+            Some(instance_id) = online_signal.recv() => {
+                debug!(?instance_id, "Online received for");
+                if let Some(connection) = pending_online_connections.remove(&instance_id) {
+                    if wait_timer_task.is_finished() {
+                        info!(?instance_id, "Defer timeout, using current connection");
+                        break connection;
+                    }
+                    // We don't defer connection if it's the first endpoint from the pecking order
+                    if instance_id == 0 {
+                        info!("Using best connection");
+                        break connection;
+                    }
+                    info!("Deferring connection {}", instance_id);
+                    non_preferred_connections.push((instance_id, connection));
+                } else {
+                    warn!(?instance_id, "Cannot find LightwayConnection");
+                }
+            },
+
+            (instance_id, outside_io_result) = mobile::lightway::first_outside_io_exit(&mut pending_online_connections) => {
+                // outside_io_task should not early return here, we're not going to use this connection anyway, removing it for clean-up
+                let connection = pending_online_connections.remove(&instance_id);
+                if let Some(connection) = connection {
+                    let _ = connection.conn.lock().unwrap().disconnect();
+                    drop(connection);
+                }
+                match outside_io_result {
+                    Ok(Err(e)) if matches!(e.downcast_ref::<ConnectionError>(), Some(ConnectionError::Unauthorized)) => {
+                        error!(?instance_id, "Unauthorized connection");
+                        connection_error_to_return = Some(ConnectionError::Unauthorized.into());
+                    }
+                    _ => error!(?instance_id, "Unexpected outside_io_task early exit: {:?}", outside_io_result),
+                }
+                failed_connections += 1;
+            }
+        }
+    };
+
+    // Drop the receiver so that no more connections can be active
+    drop(online_signal);
+
+    debug!(?active_connection.instance_id, "Using connection");
+    if let Err(e) = connected_index.set(active_connection.instance_id) {
+        warn!(
+            "Connection index has been set already, should only be called once: {}",
+            e
+        );
+    }
+
+    external_event_handler.handle_status_change(lightway_core::State::Online as u8);
+
+    non_preferred_connections.extend(pending_online_connections.drain());
+    drop(pending_online_connections);
+
+    let _ = in_progress_connection_abort_handles.swap_remove(active_connection.instance_id);
+    tokio::spawn(mobile::lightway::cleanup_connections(
+        in_progress_connection_abort_handles,
+        non_preferred_connections
+            .into_iter()
+            .map(|(_, connection)| connection)
+            .collect(),
+    ));
+
+    let mobile::lightway::LightwayConnection {
+        conn,
+        outside_io_task,
+        new_outside_io_sender,
+        keepalive,
+        keepalive_task,
+        keepalive_config,
+        expresslane_event_rx,
+        ..
+    } = active_connection;
+
+    // We are online listen for network changes
+    let network_change_task = tokio::spawn(handle_network_change(
+        keepalive.clone(),
+        network_change_receiver,
+        Arc::downgrade(&conn),
+        new_outside_io_sender,
+    ));
+
+    if let Some(mut expresslane_event_rx) = expresslane_event_rx {
+        // We only process Expresslane state changes after we selected the best connection
+        tokio::spawn(async move {
+            while let Some(state) = expresslane_event_rx.recv().await {
+                debug!(?state, "Expresslane State Change");
+                external_event_handler.handle_expresslane_state_change(state);
+            }
+        });
+    }
+
+    conn.lock().unwrap().app_state_mut().extended = Some(inside_io.clone());
+    let inside_io_loop: JoinHandle<uniffi::Result<()>> = tokio::spawn(inside_io_task(
+        conn.clone(),
+        inside_io,
+        tun_dns_ip,
+        keepalive,
+        keepalive_config,
+    ));
+
+    tokio::select! {
+        // Use biased selection to prioritize management commands and prevent race conditions
+        // where network_change_task exits early due to network change, dropping its mpsc sender and
+        // causing outside_io_task to throw channel errors and return wrong result to the app.
+        biased;
+        result = network_change_task => {
+            match result {
+                Ok(Ok(client_result)) => {
+                    info!("network change task result: {client_result:?}");
+                    Ok(client_result.into())
+                },
+                Ok(Err(e)) => {
+                    Err(anyhow!("error during network change: {e:?}"))
+                }
+                Err(e) => {
+                    Err(anyhow!("network change task error: {e:?}"))
+                }
+            }
+        }
+        Some(_) = keepalive_task => Err(anyhow!("Keepalive timeout")),
+        io = outside_io_task => match io {
+                Ok(Err(e)) if matches!(e.downcast_ref::<ConnectionError>(), Some(ConnectionError::Goodbye)) => {
+                    info!("Received server goodbye, returning result...");
+                    Ok(ClientResult::ServerGoodbye)
+                }
+                _ => Err(anyhow!("Outside IO loop exited: {io:?}"))
+        },
+        io = inside_io_loop => Err(anyhow!("Inside IO loop exited: {io:?}")),
+    }
 }
 
 #[cfg(test)]
